+++ frontend/src/components/subject/SubjectCreate.jsx
@@ -10,6 +10,7 @@ import Swal from "sweetalert2";
 import { useNavigate } from "react-router-dom";
 import { isEmpty } from "../../validations/InputValidations";
 import { Field, Form, Formik, ErrorMessage } from "formik";
+import { SubjectForm } from "./SubjectForm";
 
 const validateForm = (values) => {
   const errors = {};
@@ -55,27 +56,17 @@ const errorResponseAlert = (error) => {
 }
 //TODO Reformat carrer enum 
 export const SubjectCreate = () => {
-  const [
-    subjectCareerValue,
-    subjectCareerValueChangeHandler
-  ] = useState("INGENIERIA_DE_SOFTWARE");
-
-  const [period, setPeriod] = useState("SEMESTRAL")
-
-  const periodHandler = (event) => {
-    setPeriod(event.target.value);
-  }
 
   const navigate = useNavigate();
 
   const createSubjectHandler = (values) => {
     const subject = {
       name: values.name,
-      career: subjectCareerValue,
+      career: values.career,
       semester: values.semester,
       credits: values.credits,
       academicHours: values.academicHours,
-      period
+      period: values.period
     }
     saveSubject(subject)
       .then(response => succesResponseAlert(response))
@@ -84,9 +75,6 @@ export const SubjectCreate = () => {
 
   }
 
-  const selectCareerHandler = ({ value }) => {
-    subjectCareerValueChangeHandler(value);
-  }
   return (
     <Flex
       height={"100%"}
@@ -106,129 +94,12 @@ export const SubjectCreate = () => {
         justifyContent={"none"}
         alignItems={"center"}
       >
-        <Flex
-          justifyContent={"none"}
-          alignItems={"center"}
-          direction={"column"}
-          gap="20px"
-          width={"90%"}
-          className={style["create-subject-container"]}
-        >
-          <Formik
-            initialValues={{
-              name: "",
-              semester: "",
-              credits: "",
-              academicHours: "",
-            }}
-            onSubmit={createSubjectHandler}
-            validate={validateForm}
-          >
-            {({ errors, touched }) => (
-              <Form className={style.form}>
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-                  className={errors.name && touched.name ? style["form__item-error"] : style["form__item"]}
-                >
-                  <label style={{ fontSize: "20px", color: errors.name && touched.name ? "red" : "black",fontWeight:"bolder"  }}>Nombre</label>
-                  <Field name="name" />
-                  <ErrorMessage name="name" style={{ fontSize: "17px", color: "red" }} component={"small"} />
-                </Flex>
+        <SubjectForm
 
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-                  className={errors.semester && touched.semester ? style["form__item-error"] : style["form__item"]}
-                >
-                  <label style={{ fontSize: "20px", color: errors.semester && touched.semester ? "red" : "black",fontWeight:"bolder" }}>Semestre</label>
-                  <Field name="semester" type="number" />
-                  <ErrorMessage name="semester" style={{ fontSize: "17px", color: "red" }} component={"small"} />
-                </Flex>
-
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-                  className={style["form__item"]}
-                >
-                  <label style={{ fontSize: "20px",fontWeight:"bolder"  }}>Carrera</label>
-                  <Select
-                    onChange={selectCareerHandler}
-                    defaultValue={{ label: careers[0].name, value: careers[0].value }}
-                    noOptionsMessage={() => "No se encontraron carreras "}
-                    className={style.select}
-                    options={careers.map((career) => ({
-                      label: career.name,
-                      value: career.value,
-                    }))}
-                  />
-                </Flex>
-
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-                  className={errors.credits && touched.credits ? style["form__item-error"] : style["form__item"]}
-                >
-                  <label style={{ fontSize: "20px", color: errors.credits && touched.credits ? "red" : "black",fontWeight:"bolder"  }}>Créditos</label>
-                  <Field name="credits" type="number" />
-                  <ErrorMessage name="credits" style={{ fontSize: "17px", color: "red" }} component={"small"} />
-                </Flex>
-
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-                  className={errors.academicHours && touched.academicHours ? style["form__item-error"] : style["form__item"]}
-                >
-                  <label style={{ fontSize: "20px", color: errors.academicHours && touched.academicHours ? "red" : "black" ,fontWeight:"bolder" }}>No. horas de trabajo acacémico</label>
-                  <Field name="academicHours" type="number" />
-                  <ErrorMessage name="academicHours" style={{ fontSize: "17px", color: "red" }} component={"small"} />
-                </Flex>
-                <Flex
-                  direction={"column"}
-                  height={"auto"}
-                  alignItems={"none"}
-                  justifyContent={"none"}
-
-                >
-                  <label style={{ fontSize: "20px", color: errors.academicHours && touched.academicHours ? "red" : "black",fontWeight:"bolder"  }}>Periodo</label>
-
-                  <label style= { {fontSize: "20px"}}>
-                    Semestral
-                    <input type="checkbox" style={{ margin: "10px"}} value={"SEMESTRAL"} group="period" checked={period == "SEMESTRAL"} onChange={periodHandler} />
-
-                  </label>
-
-                  <label style= { {fontSize: "20px"}}>
-                    Trimestral
-                    <input type="checkbox" style={{ margin: "10px"}} value={"TRIMESTRAL"} group="period" checked={period == "TRIMESTRAL"} onChange={periodHandler}/>
-
-                  </label>
-
-                </Flex>
-
-                <Flex width>
-                  <Button inLineStyle={{ width: "120px", height: "40px", margin: "10px", backgroundColor: "blue" }}>
-                    Guardar
-                  </Button>
-                  <Button inLineStyle={{ width: "120px", height: "40px", margin: "10px" }} onClick={() => navigate("/asignaturas")}>
-                    Cancelar
-                  </Button>
-                </Flex>
-              </Form>
-            )}
-          </Formik>
-
-        </Flex>
+          onSubmit={createSubjectHandler}
+        
+        />
+     
       </Flex>
     </Flex>
   );
